pyproject.toml
[tool.poetry]
name = "contract-manager"
version = "0.1.0"
description = ""
authors = ["u.veka"]

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.4"
psycopg2-binary = "^2.9.7"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"